Build process now uses newer browserify.
Browserify was recently updated to fix a long-standing issue with relative paths. However, this broke my previous Windows fix. My new Windows fix is on a separate branch, domenic/browserify#windows. But, adapting to the changes in Browserify required some messing with the build script. Along the way, I documented a bunch of its stuff.